Added frame time, frame limiting, and frame rate logging

+++ src/App.h
@@ -3,5 +3,5 @@ void Init();
 void Loop();
 void Quit();
 
-float FrameTime();
-float TotalTime();
+double FrameTime();
+double TotalTime();
